chore: restructure project to allow for other packages FUI-1484
