+++ src/custom-elements-plugin/completions/fast.test.ts
