+++ src/custom-elements-plugin/completions/fast.ts
