+++ src/custom-elements-plugin/completions/helpers.ts
