+++ src/custom-elements-plugin/constants/diagnostic-codes.ts
