+++ src/custom-elements-plugin/custom-elements/manifest/analyzer.ts
