+++ src/custom-elements-plugin/custom-elements/manifest/repository.ts
