+++ src/custom-elements-plugin/custom-elements/service.ts
