+++ src/custom-elements-plugin/diagnostics/diagnostics.test.ts
