+++ src/custom-elements-plugin/diagnostics/diagnostics.ts
