+++ src/custom-elements-plugin/diagnostics/diagnostics.types.ts
