+++ src/custom-elements-plugin/diagnostics/fast.test.ts
