+++ src/custom-elements-plugin/diagnostics/fast.ts
