+++ src/custom-elements-plugin/diagnostics/index.ts
