+++ src/custom-elements-plugin/global-data/data/attributes.ts
