+++ src/custom-elements-plugin/global-data/data/events.ts
