+++ src/custom-elements-plugin/global-data/data/html-attributes.ts
