+++ src/custom-elements-plugin/global-data/data/tagnames.ts
