+++ src/custom-elements-plugin/global-data/global-data.types.ts
