+++ src/custom-elements-plugin/global-data/repository.ts
