+++ src/custom-elements-plugin/global-data/service.ts
