+++ src/custom-elements-plugin/index.ts
