+++ src/custom-elements-plugin/metadata/fast.test.ts
