+++ src/custom-elements-plugin/metadata/fast.ts
