+++ src/custom-elements-plugin/metadata/helpers.ts
