+++ src/custom-elements-plugin/metadata/index.ts
