+++ src/custom-elements-plugin/metadata/metadata.test.ts
