+++ src/custom-elements-plugin/metadata/metadata.ts
