+++ src/custom-elements-plugin/utils/index.ts
