+++ src/custom-elements-plugin/utils/kvstore.ts
