+++ src/custom-elements-plugin/utils/strings.test.ts
