+++ src/custom-elements-plugin/utils/strings.ts
